+++ features/fixtures/ios/Scenarios/AutoInstrumentNetworkCallbackScenario.swift
@@ -10,16 +10,11 @@ import Foundation
 @objcMembers
 class AutoInstrumentNetworkCallbackScenario: Scenario {
 
-    lazy var baseURL: URL = {
-        var components = URLComponents(string: Fixture.mazeRunnerURL)!
-        components.port = 9340 // `/reflect` listens on a different port :-((
-        return components.url!
-    }()
-
     override func configure() {
         super.configure()
         config.autoInstrumentNetworkRequests = true
         config.networkRequestCallback = { (info: BugsnagPerformanceNetworkRequestInfo) -> BugsnagPerformanceNetworkRequestInfo in
+            super.ignoreInternalRequests(info: info)
 
             let testUrl = info.url
             if (testUrl == nil) {