+++ CHANGELOG.md
@@ -1,12 +1,63 @@
 Changelog
 =========
 
-## TBD
+## 1.4.0 (2024-01-31)
+
+### Enhancements
+
+* swizzleViewLoadPreMain setting is now false by default. When this config value is true, all custom view controllers are swizzled at app start, which may delay app start if there are many of them.
+  [202](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/202)
+
+* Detect pre-loaded views and correct their durations. These can occur in container views, such as `UITabViewController` when the view is loaded in anticipation of the user opening the next tab. The resulting spans are now marked with a "pre-loaded" suffix and the timings adjusted to start from `viewWillAppear` only. 
+  [236](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/236)
+
+### Bug fixes
+
+* Use recursive mutexes in places that might be reentrantly accessed.
+  [242](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/242)
+
+* Fix for potential incorrect span list size due to race condition.
+  [238](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/238)
+
+* Ensure that swizzled method return values are always propagated correctly.
+  [202](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/202)
+
+## 1.3.0 (2024-01-04)
+
+### Enhancements
+
+* Added support for deferring view load span end
+  [230](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/230)
+
+* Discard unfinished spans when the app goes into the background
+  [228](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/228)
+
+## 1.2.0 (2023-12-06)
+
+This release increases the minimum supported iOS version of the library from 11 to 13.
+
+### Enhancements
+
+* Added support for instrumenting SwiftUI views
+  [222](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/222)
+
+## 1.1.3 (2023-11-23)
 
 ### Enhancements
 
 * Added [privacy manifest](https://developer.apple.com/documentation/bundleresources/privacy_manifest_files) to declare data use and required reasons for API usage
   [212](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/212)
+
+* Detect app prewarming and discard any view load spans that would be distorted by it. 
+  [211](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/211)
+
+### Bug fixes
+
+* Fetching of network swizzle targets is now done on a BG queue in order to avoid a potential deadlock 
+  [218](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/218)
+
+* Ensure span cancellation is done with concurrency protection 
+  [217](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/217)
 
 ## 1.1.2 (2023-10-19)